Add in changes from philipc@lineo:
Fix three bugs and bring into line with glibc:

1. The first character read using getc() was being ignored if it was EOF.
Normally this is okay because the next getc() returns EOF as well, but
for sscanf, this was causing us to skip the null terminator and start
scanning whatever happened to be next in memory.

2. %s, %c, and %[ formats now return -1 if EOF is reached before any
characters are read, instead of 0.  This was causing an infinite loop
in diald.

3. Default to base 10 for %i fields if not prefix modifier present.


git-svn-id: svn://uclibc.org/trunk/uClibc@3230 69ca8d6d-28ef-0310-b511-8ec308f3f277