Disable macro implementations.  Somehow I missed the following from the ANSI/ISO C99 std...
A library macro that masks a function declaration expands to an
expression that evaluates each of its arguments once (and only once).
               ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
Arguments that have side effects evaluate the same way whether the
expression executes the macro expansion or calls the function. Macros
for the functions getc and putc are explicit exceptions to this rule.
Their stream arguments can be evaluated more than once. Avoid argument
expressions that have side effects with these macros.

This will be reworked before the next release.  And yes, I feel stupid.


git-svn-id: svn://uclibc.org/trunk/uClibc@5290 69ca8d6d-28ef-0310-b511-8ec308f3f277