Patch from Jeffrey Damick:
With glibc if you do a res_init() that forces a reread of the nameservers file,
but in uClibc the res_init will not force a reread if it has already read the
file because then the nameservers value is greater than 0.

So res_init must call close_nameservers() before calling open_nameservers().


git-svn-id: svn://uclibc.org/trunk/uClibc@6656 69ca8d6d-28ef-0310-b511-8ec308f3f277