Patch from Paul Mundt <lethal@linux-sh.org>:
Here's the patch for the ldso bits for sh64. This is still in need of a bunch
of debugging, testing, etc. and is really only being submitted for general
completeness. This assumes that the previous patches I've submitted have
already been applied.

I plan on playing with this and buildroot some more later, as I'd definitely
like to see buildroot images for sh64.


git-svn-id: svn://uclibc.org/trunk/uClibc@8113 69ca8d6d-28ef-0310-b511-8ec308f3f277