Alexandre Oliva writes:
While testing the FR-V code with GCC mainline, I ran into some
problems in the RPC code.  It relies on a GCC extension that is no
longer available, namely, the result of a cast is no longer considered
an lvalue.

This patch enables the code to compile.  I haven't been able to test
RPC though, especially in a multi-threaded environment.


git-svn-id: svn://uclibc.org/trunk/uClibc@8504 69ca8d6d-28ef-0310-b511-8ec308f3f277