Joakim Tjernlund writes:
Hi Erik

It seems to me that __pthread_once and __pthread_initialize_minimal could be made
WEAKs with no stub. The code in rpc_thread.c and __uClibc_main.c appears to expect this.
Also, __pthread_return_0 __pthread_return_1 and __pthread_return_void can be static, not to
pollute the name space.

  Jocke


git-svn-id: svn://uclibc.org/trunk/uClibc@8841 69ca8d6d-28ef-0310-b511-8ec308f3f277