Stephen Hemminger from osdl dot org writes:
Rather than copy more code back into the bridge-utilities, how about applying this
change to uClibc?  I hate when packages get cluttered for workarounds for other
incompatibilities.

The problem is that SIOCGIFCONF only lists interfaces that have IP addresses, so it
doesn't find the other interfaces that are being used for bridging. It could be fixed
in the kernel to return all interfaces, but then something else might break; and still
it mean a kernel update for the 2.4 users.

The whole use of ifindex in the bridge API is a bad idea. But we probably have to live
with it for compatibility.

Patch against uClibc 0.9.26


git-svn-id: svn://uclibc.org/trunk/uClibc@8894 69ca8d6d-28ef-0310-b511-8ec308f3f277