Joakim Tjernlund writes:
Hi yet again :)

in dl-startup.c when performing boot strap relocation the following test
exists to make sure that only "_dl_" symbols are relocated:
/* We only do a partial dynamic linking right now.  The user
 is not supposed to define any symbols that start with a
 '_dl', so we can do this with confidence. */
 if (!symname || !_dl_symbol(symname)) {
        continue;
 }

However on PPC(and the other archs as well I suspect) all symbols are
"_dl_" symbols so the test is never true. The test can be removed and the
whole loop simplified(smaller). This also makes it possible to
simplify elfinterp.c

This remove the scanning of ldso.so relocs, making relocation faster.

I have tested this on PPC and it works well.
Do you think this optimization will work for the other arches as well?
I can't see why not.

     Jocke

* Tested on x86, arm, mipsel, and powerpc by Erik and works nicely
 -Erik


git-svn-id: svn://uclibc.org/trunk/uClibc@8930 69ca8d6d-28ef-0310-b511-8ec308f3f277