Pete Popov writes:
Hi Erik,

I'm not sure why the NIOS support is not in uClibc -- perhaps the patch
was rejected or never submitted? In any case, I'm playing with some NIOS
stuff and created this patch against 0.9.26. The work was done by
Microtronix. I'm not sure who else contributed to it. It would be great
to have the NIOS support available in uClibc so developers don't have to
go searching for these bits.

Pete


git-svn-id: svn://uclibc.org/trunk/uClibc@8957 69ca8d6d-28ef-0310-b511-8ec308f3f277