Hi Erik and Manuel
Long time no see :)

It appears uClibc pthreads native debugging is broken w.r.t thread exit handling in
uClibc(at least on PPC). When debugging ex7, gdb(6.2.1) bails out as
soon as a thread exits. I found a comment in gdb that TD_DEATH handling was broken for
glibc 2.3.1, so I figured that maybe it was broken in uClibc also.

I added a #if 0 #endif in pthread_exit( see patch below) and then
gdb behaved again.

While looking into this I found a few differences( included in the patch)
between glibc and uClibc. Don't know if these makes a difference, but I leave that to
you to decide.

 Jocke


git-svn-id: svn://uclibc.org/trunk/uClibc@9785 69ca8d6d-28ef-0310-b511-8ec308f3f277