John Bowler writes in Bug 385: This patch changes all cases where the ARM assembler mov pc,rx instructions are used to ensure that the thumb/arm interwork change of process more works - in essence mov pc,rx needs to become bx rc.
The ldr pc or ldm rx, {pc} instructions are not changed - this is
fine on ARM >=v5 but will fail to restore thumb mode on ARM v4T,
i.e. this code will not provide support for thumb on ARM v4T.

One mov pc is left in resolve.S, this is fixed in a different patch -
thumb-resolve.patch

The changes are protected by __THUMB_INTERWORK__ - the original
mov instruction will work on newer architectures and is required on
arch v4 (not v4t) and earlier - those which did not support thumb -
so this is safe.  See gcc lib1asmfuncs for a more exact test.


git-svn-id: svn://uclibc.org/trunk/uClibc@13735 69ca8d6d-28ef-0310-b511-8ec308f3f277