Richard Sandiford writes: Until a few months ago, libc/sysdeps/linux/m68k/clone.S defined __syscall_error itself.  This local definition was removed in revision 13305:
However, the clone code uses bcc.w to branch to __syscall_error.
This made sense when there was a local definition, because the
label was guaranteed to be in range.  We can't guarantee that
the external definition will be in range though; it depends on
the order that the objects are linked.

This patch adjusts clone.S to use jbcc instead.


git-svn-id: svn://uclibc.org/trunk/uClibc@15309 69ca8d6d-28ef-0310-b511-8ec308f3f277